Move generic content action and associated functions into a generic content controller
